Add eldap to base-local precise
This commit adds the erlang-eldap package to the precise
base-local image. This is needed to support the ldap_auth app
which is now included in Local.

Note we do not need to make a similar change for the squeeze
base-local image because erlang is obtained via a tarball which
already includes eldap.

BugzID: 58073